requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
click==8.3.1
ecdsa==0.19.2
fastapi==0.125.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.3
pydantic==1.10.26
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing_extensions==4.15.0
uvicorn==0.41.0
